feat: logs viewer v3 - path routes, expand-in-URL, infinite scroll
Refactor the logs viewer to use Nuxt nested routes:

- /developer/logs/events and /developer/logs/spans become real paths
  instead of a '?view=' query param. The bare /developer/logs route
  redirects to /events. The tabs in the toolbar preserve the rest of
  the query string (filters + expand state) when switching tabs.
- Inline structured fields is on by default and now also applies to
  span rows on the Spans tab.
- Expanded event / span rows are written to the URL as ?expand=e-N and
  ?expand=s-N. The same param carries both kinds, prefixed to avoid id
  collisions between the events and spans tables. Both lists persist
  across tab switches.
- Event and span rows now render the date alongside the time, using
  a toLocaleString('short month, day, h:mm:ss') format.
- Infinite scroll switched from a hand-rolled IntersectionObserver to
  @vueuse/core's useInfiniteScroll attached to UScrollArea's root
  element. The page now scrolls inside UScrollArea (the dashboard panel
  body is set to overflow-hidden so there's no double scrollbar) and
  emits loading more rows as the user scrolls near the bottom.

Shared state lives in a new useLogsContext composable and is provided
from logs.vue to its child pages via inject.

.gitignore's bare 'logs' rule was shadowing the new app/pages/developer/
logs/ directory; anchored it to /logs so route folders work.

OpenAPI refreshed: the listLogs q param is documented as a substring
search across message, target, and structured fields (it used to claim
a now-removed full-text search on the message column only). boot_id is
now formatted as uuid in the version and boots responses.
