+++ .gitignore
@@ -13,7 +13,7 @@ dist
 node_modules
 
 # Logs
-logs
+/logs
 *.log
 
 # Misc
