+++ app/pages/developer/logs/index.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+definePageMeta({
+  middleware: [
+    () => {
+      const localePath = useLocalePath();
+      return navigateTo(localePath("/developer/logs/events"));
+    },
+  ],
+});
+</script>
+
+<template>
+  <div />
+</template>
