+++ openapi/openapi.json
@@ -392,7 +392,7 @@
           {
             "name": "q",
             "in": "query",
-            "description": "Full-text search on the message field.",
+            "description": "Substring search across message, target, and structured fields.",
             "required": false,
             "schema": {
               "type": "string"
@@ -785,6 +785,7 @@
         "properties": {
           "boot_id": {
             "type": "string",
+            "format": "uuid",
             "description": "Unique boot id (UUID)."
           },
           "event_count": {
@@ -1564,6 +1565,7 @@
           },
           "boot_id": {
             "type": "string",
+            "format": "uuid",
             "description": "Unique id of this gateway boot session."
           }
         }
